+++ projects/orquestra-workflow-runtime/tests/test_build_workflow.py
@@ -1,7 +1,3 @@
-################################################################################
-# © Copyright 2024 Zapata Computing Inc.
-################################################################################
-
 import re
 from typing import Any, Dict, List, Optional, Union
 from unittest.mock import ANY, Mock, call, create_autospec
@@ -15,11 +11,6 @@ from orquestra.workflow_shared.exceptions import OrquestraSDKVersionMismatchWarn
 from orquestra.workflow_shared.schema import ir
 from orquestra.workflow_shared.schema.ir import GitURL, SecretNode
 from orquestra.workflow_shared.schema.responses import WorkflowResult
-
-import orquestra.sdk as sdk
-from orquestra.sdk._client._base._testing._example_wfs import (
-    workflow_parametrised_with_resources,
-)
 
 
 @pytest.fixture
@@ -122,21 +113,6 @@ class TestBuildGitURL:
         with pytest.raises(ValueError) as exc_info:
             _ = _build_workflow._build_git_url(git_url, "custom_protocol")
         exc_info.match("Unknown protocol: `custom_protocol`")
-
-
-def make_workflow_with_dependencies(deps, *, n_tasks=1):
-    """Generate a workflow definition with the specified dependencies."""
-
-    @sdk.task(dependency_imports=deps)
-    def hello_orquestra() -> str:
-        return "Hello Orquestra!"
-
-    @sdk.workflow()
-    def hello_orquestra_wf():
-        return [hello_orquestra() for _ in range(n_tasks)]
-
-    return hello_orquestra_wf()
-
 
 class TestPipString:
     @pytest.fixture()
@@ -303,191 +279,6 @@ class TestPipString:
             assert pip == []
 
 
-class TestMakeDag:
-    @pytest.fixture
-    def client(self):
-        return create_autospec(_client.RayClient)
-
-    @pytest.fixture
-    def wf_run_id(self):
-        return "mocked_wf_run_id"
-
-    def test_import_pip_string_resolved_once_per_import(
-        self, monkeypatch: pytest.MonkeyPatch, client: Mock, wf_run_id: str
-    ):
-        pip_string = create_autospec(_build_workflow._pip_string)
-        pip_string.return_value = ["mocked"]
-        monkeypatch.setattr(_build_workflow, "_pip_string", pip_string)
-        imps = [
-            sdk.GithubImport(
-                "zapata-engineering/orquestra-sdk",
-                personal_access_token=sdk.Secret(
-                    "mock-secret", config_name="mock-config", workspace_id="mock-ws"
-                ),
-            ),
-            sdk.PythonImports("numpy", "polars"),
-            sdk.InlineImport(),
-        ]
-        workflow_def = make_workflow_with_dependencies(imps, n_tasks=100).model
-        _ = _build_workflow.make_ray_dag(client, workflow_def, wf_run_id, False)
-        assert pip_string.mock_calls == [
-            call(imp) for imp in workflow_def.imports.values()
-        ]
-
-    class TestResourcesInMakeDag:
-        @pytest.mark.parametrize(
-            "resources, expected, types",
-            [
-                ({}, {}, {}),
-                ({"cpu": "1000m"}, {"num_cpus": 1.0}, {"num_cpus": int}),
-                ({"memory": "1Gi"}, {"memory": 1073741824}, {"memory": int}),
-                ({"gpu": "1"}, {"num_gpus": 1}, {"num_gpus": int}),
-                (
-                    {"cpu": "2500m", "memory": "10G", "gpu": "1"},
-                    {"num_cpus": 2.5, "memory": 10000000000, "num_gpus": 1},
-                    {"num_cpus": float, "memory": int, "num_gpus": int},
-                ),
-            ],
-        )
-        def test_setting_resources(
-            self,
-            client: Mock,
-            wf_run_id: str,
-            resources: Dict[str, str],
-            expected: Dict[str, Union[int, float]],
-            types: Dict[str, type],
-        ):
-            # Given
-            workflow = workflow_parametrised_with_resources(**resources).model
-
-            # When
-            _ = _build_workflow.make_ray_dag(client, workflow, wf_run_id, False)
-
-            # Then
-            calls = client.add_options.call_args_list
-
-            # We should only have two calls: our invocation and the aggregation step
-            assert len(calls) == 2
-            # Checking our call did not have any resources included
-            assert calls[0] == call(
-                ANY,
-                name=ANY,
-                metadata=ANY,
-                runtime_env=ANY,
-                catch_exceptions=ANY,
-                max_retries=ANY,
-                **expected,
-            )
-            for kwarg_name, type_ in types.items():
-                assert isinstance(calls[0].kwargs[kwarg_name], type_)
-
-        @pytest.mark.parametrize(
-            "custom_image, gpu, expected_resources, expected_kwargs",
-            (
-                (
-                    "a_custom_image:latest",
-                    None,
-                    {"image:a_custom_image:latest": 1},
-                    {},
-                ),
-                (
-                    None,
-                    None,
-                    {
-                        "image:hub.nexus.orquestra.io/zapatacomputing/orquestra-sdk-base:mocked": 1  # noqa: E501
-                    },
-                    {},
-                ),
-                (
-                    None,
-                    1,
-                    {
-                        "image:hub.nexus.orquestra.io/zapatacomputing/orquestra-sdk-base:mocked-cuda": 1  # noqa: E501
-                    },
-                    {
-                        "num_gpus": 1,
-                    },
-                ),
-            ),
-        )
-        class TestSettingCustomImage:
-            def test_with_env_set(
-                self,
-                client: Mock,
-                wf_run_id: str,
-                monkeypatch: pytest.MonkeyPatch,
-                custom_image: Optional[str],
-                gpu: Optional[int],
-                expected_resources: Dict[str, int],
-                expected_kwargs: Dict[str, Any],
-            ):
-                # Given
-                monkeypatch.setenv("ORQ_RAY_SET_CUSTOM_IMAGE_RESOURCES", "1")
-                workflow = workflow_parametrised_with_resources(
-                    gpu=gpu, custom_image=custom_image
-                ).model
-
-                # To prevent hardcoding a version number, let's override the version for
-                # this test.
-
-                # We can be certain the workfloe def metadata is available
-                assert workflow.metadata is not None
-                workflow.metadata.sdk_version.original = "mocked"
-
-                # When
-                _ = _build_workflow.make_ray_dag(client, workflow, wf_run_id, False)
-
-                # Then
-                calls = client.add_options.call_args_list
-
-                # We should only have two calls: our invocation and the aggregation step
-                assert len(calls) == 2
-                # Checking our call did not have any resources included
-                assert calls[0] == call(
-                    ANY,
-                    name=ANY,
-                    metadata=ANY,
-                    runtime_env=ANY,
-                    catch_exceptions=ANY,
-                    max_retries=ANY,
-                    resources=expected_resources,
-                    **expected_kwargs,
-                )
-
-            def test_with_env_not_set(
-                self,
-                client: Mock,
-                wf_run_id: str,
-                custom_image: Optional[str],
-                gpu: Optional[int],
-                expected_resources: Dict[str, int],
-                expected_kwargs: Dict[str, Any],
-            ):
-                # Given
-                workflow = workflow_parametrised_with_resources(
-                    gpu=gpu, custom_image=custom_image
-                ).model
-
-                # When
-                _ = _build_workflow.make_ray_dag(client, workflow, wf_run_id, False)
-
-                # Then
-                calls = client.add_options.call_args_list
-
-                # We should only have two calls: our invocation and the aggregation step
-                assert len(calls) == 2
-                # Checking our call did not have any resources included
-                assert calls[0] == call(
-                    ANY,
-                    name=ANY,
-                    metadata=ANY,
-                    runtime_env=ANY,
-                    catch_exceptions=ANY,
-                    max_retries=ANY,
-                    **expected_kwargs,
-                )
-
-
 class TestArgumentUnwrapper:
     @pytest.fixture
     def mock_secret_get(self, monkeypatch: pytest.MonkeyPatch):
@@ -643,178 +434,3 @@ class TestArgumentUnwrapper:
 
             calls = [call(packed), call(unpacked[0]), call(unpacked[1])]
             mock_deserialize.assert_has_calls(calls)
-
-
-@pytest.mark.parametrize(
-    "installed_sdk_version, expected_sdk_dependency",
-    [
-        ("0.57.1.dev3+g3ef9f57.d20231003", None),
-        ("1.2.3", "1.2.3"),
-        ("0.1.dev1+g25df81e", None),
-    ],
-)
-class TestHandlingSDKVersions:
-    """``_import_pip_env`` handles adding the current SDK version as a dependency.
-
-    Note that these tests don't mock serde - we're interested in whether the correct
-    package list gets constructed so we can't just return 'mocked' for imports.
-    """
-
-    @staticmethod
-    def test_with_no_dependencies(
-        installed_sdk_version: str,
-        expected_sdk_dependency: str,
-        monkeypatch: pytest.MonkeyPatch,
-    ):
-        # Given
-        monkeypatch.setattr(
-            _build_workflow,
-            "get_installed_version",
-            Mock(return_value=installed_sdk_version),
-        )
-        wf = make_workflow_with_dependencies([]).model
-        task_inv = [inv for inv in iter_invocations_topologically(wf)][0]
-        imports = {
-            id: _build_workflow._pip_string(imp) for id, imp in wf.imports.items()
-        }
-
-        # When
-        pip = _build_workflow._import_pip_env(task_inv, wf, imports)
-
-        # Then
-        if expected_sdk_dependency:
-            assert pip == [f"orquestra-sdk=={expected_sdk_dependency}"]
-        else:
-            assert pip == []
-
-    @staticmethod
-    @pytest.mark.parametrize(
-        "python_imports",
-        [
-            ["MarkupSafe==1.0.0"],
-            ["MarkupSafe==1.0.0", "Jinja2==2.7.2"],
-        ],
-    )
-    def test_with_multiple_dependencies(
-        python_imports: List[str],
-        installed_sdk_version: str,
-        expected_sdk_dependency: str,
-        monkeypatch: pytest.MonkeyPatch,
-    ):
-        # Given
-        monkeypatch.setattr(
-            _build_workflow,
-            "get_installed_version",
-            Mock(return_value=installed_sdk_version),
-        )
-        wf = make_workflow_with_dependencies([sdk.PythonImports(*python_imports)]).model
-        task_inv = [inv for inv in iter_invocations_topologically(wf)][0]
-        imports = {
-            id: _build_workflow._pip_string(imp) for id, imp in wf.imports.items()
-        }
-
-        # When
-        pip = _build_workflow._import_pip_env(task_inv, wf, imports)
-
-        # Then
-        if expected_sdk_dependency:
-            assert sorted(pip) == sorted(
-                python_imports + [f"orquestra-sdk=={expected_sdk_dependency}"]
-            )
-        else:
-            assert sorted(pip) == sorted(python_imports)
-
-    @pytest.mark.parametrize(
-        "sdk_import",
-        [
-            "orquestra-sdk",
-            "orquestra-sdk>=1.2.3",
-            "orquestra-sdk<=1.2.3",
-            "orquestra-sdk~=1.2.3",
-            "orquestra-sdk==1.2.3",
-            "orquestra-sdk!=1.2.3",
-            "orquestra-sdk===1.2.3",
-            "orquestra-sdk >= 1.2.3",
-            "orquestra-sdk <= 1.2.3",
-            "orquestra-sdk ~= 1.2.3",
-            "orquestra-sdk == 1.2.3",
-            "orquestra-sdk != 1.2.3",
-            "orquestra-sdk === 1.2.3",
-        ],
-    )
-    @pytest.mark.parametrize(
-        "python_imports",
-        [
-            [],
-            ["MarkupSafe==1.0.0"],
-            ["MarkupSafe==1.0.0", "Jinja2==2.7.2"],
-        ],
-    )
-    class TestHandlesSDKDependency:
-        @staticmethod
-        @pytest.mark.filterwarnings("ignore:The definition for task ")
-        def test_replaces_declared_sdk_dependency(
-            sdk_import,
-            python_imports,
-            installed_sdk_version: str,
-            expected_sdk_dependency: str,
-            monkeypatch: pytest.MonkeyPatch,
-        ):
-            # Given
-            monkeypatch.setattr(
-                _build_workflow,
-                "get_installed_version",
-                Mock(return_value=installed_sdk_version),
-            )
-            wf = make_workflow_with_dependencies(
-                [sdk.PythonImports(*python_imports + [sdk_import])]
-            ).model
-            task_inv = [inv for inv in iter_invocations_topologically(wf)][0]
-            imports = {
-                id: _build_workflow._pip_string(imp) for id, imp in wf.imports.items()
-            }
-
-            # When
-            pip = _build_workflow._import_pip_env(task_inv, wf, imports)
-
-            # Then
-            if expected_sdk_dependency:
-                assert sorted(pip) == sorted(
-                    python_imports + [f"orquestra-sdk=={expected_sdk_dependency}"]
-                )
-            else:
-                assert sorted(pip) == sorted(python_imports)
-
-        @staticmethod
-        @pytest.mark.filterwarnings("error")
-        def test_warns_user_that_declared_sdk_dependency_is_ignored(
-            sdk_import,
-            python_imports,
-            installed_sdk_version: str,
-            expected_sdk_dependency: str,
-            monkeypatch: pytest.MonkeyPatch,
-        ):
-            # Given
-            monkeypatch.setattr(
-                _build_workflow,
-                "get_installed_version",
-                Mock(return_value=installed_sdk_version),
-            )
-            wf = make_workflow_with_dependencies(
-                [sdk.PythonImports(*python_imports + [sdk_import])]
-            ).model
-            task_inv = [inv for inv in iter_invocations_topologically(wf)][0]
-            imports = {
-                id: _build_workflow._pip_string(imp) for id, imp in wf.imports.items()
-            }
-
-            # When
-            with pytest.raises(OrquestraSDKVersionMismatchWarning) as e:
-                _ = _build_workflow._import_pip_env(task_inv, wf, imports)
-
-            # Then
-            warning: str = e.exconly().strip()
-            assert re.match(
-                r"^orquestra\.workflow_shared\.exceptions\.OrquestraSDKVersionMismatchWarning: The definition for task `task-hello-orquestra-.*` declares `orquestra-sdk(?P<dependency>.*)` as a dependency. The current SDK version (\((?P<installed>.*)\) )?is automatically installed in task environments. The specified dependency will be ignored.$",  # noqa: E501
-                warning,
-            ), warning
