--- projects/orquestra-workflow-runtime/tests/test_fake.py
@@ -1,6 +1,0 @@
-# THIS IS JUST A FAKE TEST TO KEEP TESTS DIRECTORY ON GIT.
-# TO BE REMOVED AFTER REAL TESTS ARE PROPERLY PORTED TO SUBPACKAGES
-
-
-def test_pass():
-    return
